Move `Pulse` and `Latch` to `Prim.Mid.*` hierarchy
