+++ reactive-banana/src/Reactive/Banana/Prim/Mid.hs
@@ -41,12 +41,12 @@ module Reactive.Banana.Prim.Mid (
 
 
 import Control.Monad.IO.Class
-import Reactive.Banana.Prim.Low.Compile
-import Reactive.Banana.Prim.Low.IO
-import Reactive.Banana.Prim.Low.Plumbing
-    ( neverP, alwaysP, liftBuild, buildLater, buildLaterReadNow, liftIOLater )
-import Reactive.Banana.Prim.Low.Types
 import Reactive.Banana.Prim.Mid.Combinators
+import Reactive.Banana.Prim.Mid.Compile
+import Reactive.Banana.Prim.Mid.IO
+import Reactive.Banana.Prim.Mid.Plumbing
+    ( neverP, alwaysP, liftBuild, buildLater, buildLaterReadNow, liftIOLater )
+import Reactive.Banana.Prim.Mid.Types
 import Reactive.Banana.Prim.High.Cached
 
 {-----------------------------------------------------------------------------
